requirements.txt
fastapi
uvicorn
pydantic-settings
supabase
pydantic-ai
aiofiles
python-dotenv
httpx
tiktoken
pymupdf
pdfminer.six
datasketch
tenacity